FFMS2: Simplify the avisynth filters Small api change => you can now list a group of desired colorspace output formats for automatic selection, less functions can return errors Less singnedness warnings
Originally committed to SVN as r2978.

+++ aegisub/FFmpegSource2/ffvideosource.h
@@ -69,7 +69,7 @@ public:
 	FFTrack *GetFFTrack() { return &Frames; }
 	virtual TAVFrameLite *GetFrame(int n, char *ErrorMsg, unsigned MsgSize) = 0;
 	TAVFrameLite *GetFrameByTime(double Time, char *ErrorMsg, unsigned MsgSize);
-	int SetOutputFormat(int TargetFormats, int Width, int Height, char *ErrorMsg, unsigned MsgSize);
+	int SetOutputFormat(int64_t TargetFormats, int Width, int Height, char *ErrorMsg, unsigned MsgSize);
 	void ResetOutputFormat();
 };
